requirements.txt
openai==1.82.0
openpyxl==3.1.5
python-dotenv==1.1.0
